requirements.txt
discord-py-slash-command==2.3.2
discord.py==1.6.0
python-dotenv==0.19.0